2023/10/28 新ブログサイトがほぼ完成しました。

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use App\Models\Group;
+use App\Models\Blog;
 
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\Request;
@@ -17,58 +18,116 @@ class CategoryController extends Controller
         return response()->json($categories);
     }
 
-    public function getAllCategories ()
-    {
+    private function getCurrent_group($id) {
+        $group = Group::find($id);
+        if(!$group) {
+            return response()->json([
+                'message' => 'そのグループは存在しません。',
+            ], 404);
+        }
 
-        if(auth()->user()->role_id == 2) {
-            $group = Group::find(auth()->user()->group_id);
-            if($group) {
-                $categories = $group->categories()->get();
-                return response()->json($categories);
+        if(auth()->user()->role_id != 1) {
+
+            $allowed_categories = json_decode(auth()->user()->allowed_categories);
+
+            $allowed = false;
+            foreach ($allowed_categories as $key => $categories) {
+                if($key == $id) {
+                    if(count($categories) > 0) {
+                        $allowed  = true; 
+                        break;
+                    }
+                }
+            }
+
+            if(auth()->user()->role_id == 2) {
+                
+                if(!$allowed && $id != '1' &&  $id != auth()->user()->group_id) {
+                    return response()->json([
+                        'message' => 'そのグループへのアクセス権がありません。',
+                    ], 400);
+                }
             } else {
-                return response()->json([
-                    "message" => "グループは存在しません",
-                ], 404);
+                if(!$allowed) {
+                    return response()->json([
+                        'message' => 'そのグループへのアクセス権がありません。',
+                    ], 400);
+                }
             }
         }
+
+        $categories = $group->categories()->get();
+        $temp = [];
+        foreach ($categories as $key => $category) {
+            $blogs = Blog::where('category_id', $category->id)->get();
+            $c = count($blogs);
+            $category['blog_count'] = $c;
+        }
+
+        $blogs = Blog::where('group_id', $group->id)->get();
+        $group['blog_count'] = count($blogs);
+        $group['categories'] = $categories;
+
+        return $group;
+    }
+
+
+    public function getCurrentGroup ($id) {
         
+        $group = $this->getCurrent_group($id);
+        return response()->json($group);
+    }
+
+
+    //super admin
+    public function getAllGroup ()
+    {
         $groups = Group::all();
-        $temp = array();
+
+        $all_categories = [];
         foreach ($groups as $key => $group) {
-            $group_temp  = array(
-                "id" => $group->id,
-                "name" => $group->name,
-                "categories"=> []
+            $categories = $group->categories()->get();
+            $temp = array(
+                $group->id => $categories
             );
 
-            $categories = $group->categories()->get();
-            foreach ($categories as $key1 => $category) {
-                $group_temp['categories'][]  = $category;
-            }
-            $temp[] = $group_temp;           
+            $all_categories[$group->id] = $categories;
         }
 
-        return response()->json($temp);
-     
+        return response()->json([
+            "groups" => $groups,
+            "categories" => $all_categories
+        ]);
     }
+
 
     public function addCategory (Request $request) 
     {
         $validator = Validator::make($request->all(), [
-            'name' => 'required|string|between:2,100',
+            'name' => 'required|string|between:1,100',
+            'group_id' => 'required'
         ]);
         if($validator->fails()){
             return response()->json($validator->errors(), 400);
         }
 
+        if(auth()->user()->role_id == 2 && $request->group_id != auth()->user()->group_id)  
+        {
+            return response()->json([
+                'message' => 'あなたはこのグループに対する編集権限を持っていません。',
+            ], 400);
+        }
+ 
         $category = new Category;
         $category->name = $request->name;
-        $category->group_id = auth()->user()->role_id == 1 ? $request->group_id : auth()->user()->group_id;
+        $category->group_id = $request->group_id;
         $category->save();
 
+        $parent_group = $this->getCurrent_group($request->group_id);
+
         return response()->json([
-            'message' => 'キャティゴリが正常に作成されました。',
-            'category' => $category
+            'message' => 'カテゴリーが正常に作成されました。',
+            'current_group' => $parent_group
         ], 201);
     } 
 
@@ -79,9 +138,23 @@ class CategoryController extends Controller
         if (!$category) {
             return response()->json(['message' => 'カテゴリが見つかりません'], 404);
         }
+
+        if(auth()->user()->role_id == 2 && $request->group_id != auth()->user()->group_id)  
+        {
+            return response()->json([
+                'message' => 'あなたはこのグループに対する編集権限を持っていません。',
+            ], 400);
+        }
     
         $category->delete();
-        return response()->json(['message' => 'カテゴリが正常に削除されました'], 201);
+
+        $parent_group = $this->getCurrent_group($category->group_id);
+
+        return response()->json([
+            'message' => 'カテゴリが正常に削除されました',
+            'current_group' => $parent_group
+        ], 201);
+        // return response()->json(['message' => 'カテゴリが正常に削除されました'], 201);
 
     }
 
@@ -99,13 +172,23 @@ class CategoryController extends Controller
         if (!$category) {
             return response()->json(['message' => 'カテゴリが見つかりません'], 404);
         }
+
+        if(auth()->user()->role_id == 2 && $request->group_id != auth()->user()->group_id)  
+        {
+            return response()->json([
+                'message' => 'あなたはこのグループに対する編集権限を持っていません。',
+            ], 400);
+        }
+ 
+
         $category->name = $request->name;
         $category->update();
 
-        return response()->json(
-            [
-                'message' => 'カテゴリが正常に更新されました',
-                'category' => $category
-            ], 201);
+        $parent_group = $this->getCurrent_group($request->group_id);
+
+        return response()->json([
+            'message' => 'カテゴリが正常に更新されました',
+            'current_group' => $parent_group
+        ], 201);
     }
  }
